+++ src/app/app.component.ts
@@ -5,8 +5,9 @@ import { StatusBar, Splashscreen } from 'ionic-native';
 import { Page1 } from '../pages/page1/page1';
 import { Page2 } from '../pages/page2/page2';
 import { ProfileComponent } from '../pages/profile/profile';
-import { PlaceOrderComponent } from './place-order/place-order.component';
-import { LaundryMap } from '../pages/map/map.component'
+import { PlaceOrderComponent } from '../place-order/place-order.component';
+import { LaundryMap } from '../pages/map/map.component';
+import { LaundryItems } from '../pages/laundryitems/laundryitems';
 
 
 @Component({
@@ -28,7 +29,8 @@ export class MyApp {
       { title: 'Page 2', component: Page2 },
       { title: 'Profile', component: ProfileComponent },
       { title: 'Place Order', component: PlaceOrderComponent},
-      { title: 'Map', component: LaundryMap }
+      { title: 'Map', component: LaundryMap },
+      { title: 'LaundryItems', component: LaundryItems}
     ];
 
   }
