+++ src/app/app.module.ts
@@ -4,7 +4,7 @@ import { MyApp } from './app.component';
 import { Page1 } from '../pages/page1/page1';
 import { Page2 } from '../pages/page2/page2';
 import { ProfileComponent } from '../pages/profile/profile';
-import { PlaceOrderModule } from './place-order/place-order.module';
+import { PlaceOrderModule } from '../place-order/place-order.module';
 import { LaundryMap } from '../pages/map/map.component'
 @NgModule({
   declarations: [
