--- src/app/place-order/place-order.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'place-order',
-    templateUrl: 'place-order.template.html',
-})
-
-export class PlaceOrderComponent{
-    constructor(){}
-}
