+++ src/pages/laundryitems/laundryitems.html
@@ -1,0 +1,11 @@
+<ion-header>
+  <ion-navbar>
+    <button ion-button menuToggle>
+      <ion-icon name="menu"></ion-icon>
+    </button>
+  </ion-navbar>
+</ion-header>
+
+<ion-content padding>
+  <h3>Laundry Items</h3>
+</ion-content>
