+++ src/pages/laundryitems/laundryitems.scss
@@ -1,0 +1,3 @@
+laundryitems-page1 {
+
+}
