+++ src/pages/laundryitems/laundryitems.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+import { NavController } from 'ionic-angular';
+
+@Component ({
+    selector: 'laundry_items',
+    templateUrl: 'laundryitems.html'
+})
+
+export class LaundryItems{
+     constructor(){}
+}
