+++ src/pages/profile/profile.ts
@@ -1,10 +1,15 @@
-import { Component } from '@angular/core';
+import { Component } from '@angular/core'
+import { NavController } from 'ionic-angular';
 
 @Component({
-  selector: 'profile',
-  templateUrl: 'profile.html'
+    selector: 'profile',
+    templateUrl: 'profile.html'
 })
 
-export class ProfileComponent {
+export class ProfileComponent
+{
+   constructor(public navCtrl: NavController) {
+        
+  }
 
 }
