--- src/place-order/place-order.html
@@ -1,10 +1,0 @@
-<!-- Disable swipe-to-go-back because it's poor UX to combine STGB with side menus -->
-<!--<ion-nav [root]="rootPage" #content swipeBackEnabled="false"></ion-nav>--><ion-content padding>
-  <h3>PLace New Order Component</h3>
-
-  <p>
-    If you get lost, the <a href="http://ionicframework.com/docs/v2">docs</a> will show you the way.
-  </p>
-
-  <button ion-button secondary menuToggle>Toggle Menu</button>
-</ion-content>
