+++ src/place-order/place-order.module.ts
@@ -1,13 +1,18 @@
-import { NgModule } from '@angular/core';
+import { NgModule } from '@angular/core'
+import { IonicModule } from 'ionic-angular';
+import { PlaceOrderComponent } from './place-order.component';
+import { LaundryItems } from '../pages/laundryitems/laundryitems';
+
 
 @NgModule({
-  declarations: [
-  ],
-  imports: [
-  ],
-  entryComponents: [
-   
-  ],
-  providers: []
+    imports: [
+        IonicModule.forRoot(PlaceOrderComponent)
+    ],
+    declarations: [PlaceOrderComponent,LaundryItems],
+    entryComponents: [PlaceOrderComponent , LaundryItems],
+    exports: [PlaceOrderComponent]
 })
-export class PlaceOrder {}
+
+export class PlaceOrderModule{
+    constructor(){}
+}
